LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/moveBefore-name-map-expected.txt



FAIL moveBefore() correctly updates name map assert_equals: expected (undefined) undefined but got (object) Element node <img name="target" data-target=""></img>
